Initial code upload
First set of test programs. One generates a simple sine wave, one generates an amplitude modulated square wave, and the third generates an amplitude modulated sine wave.

Uploading a bunch of crap the IntelliJ generates along with it because I have no idea which files are necessary or not. Might as well include them all (save for the compiled stuff).