Updated maths
So after a quick read of some wiki articles and other pages online I think I have a good grasp on the basic math behind a sine wave. The updated code reflects this understanding, as it is now possible to specify the exact number of cycles shown on the screen as well as full control over the wave's phase.

Also changed the orthoview so the far left border is now 0 with the far right side WINDOW_SIZE_WIDTH long. The top and bottom planes remain at 3 since that seems to work well enough already.

It is rather trivial to add keyboard callbacks to allow live modification of the waveform so I will do that next. Although before that I will port these changes over to the amplitude modulated examples I wrote yesterday and see if that improves the atrocious-looking code.