ported changes
ported changes to other scripts. Looks ok for now. The square wave example (really a sawtooth wave but whatevs) still looks messy and terrible but mostly because I dont know how to generate a pure sawtooth wave. Would be easier to modulate then but at the moment we are kind of rigging it to work.