tiangles...
Was going to do a semicircle too but the math for the first kind of bessier function is so far over my head its hilarious. I found a formula for determining the output of a specific order bessel function but Im not sure if its exactly what I need.  Will have to look into gamma functions as well. :(