Fixed fourier-series signals phaseshift/frequency
Phase shift and frequency modifications were not actually working for the fourier-series based signals. I tried to insert the changes for phaseshifting in the actual maths for those signals but it doesnt work. Increasing frequency just depends on changing L. For phaseshifting I decided to just cheat and move the entire drawing space instead of messing with the signals. It works perfectly so I'm ok with it.